+++ resources/views/admin/positions/index.blade.php
@@ -15,11 +15,13 @@
                                     <th>#</th>
                                     <th>{{ __('Position') }}</th>
                                     <th>{{ __('Type') }}</th>
-                                    <th>
-                                        <a href="{{ url('admin/positions/create') }}" class="btn m-btn--pill m-btn--air btn-secondary table-middle " data-toggle="m-tooltip" data-placement="left" data-original-title="{{ __('Add new') }}">
-                                            <i class="flaticon-add"></i>
-                                        </a>
-                                    </th>
+                                    @if (Auth::user()->role == config('site.permission.admin'))
+                                        <th>
+                                            <a href="{{ url('admin/positions/create') }}" class="btn m-btn--pill m-btn--air btn-secondary table-middle " data-toggle="m-tooltip" data-placement="left" data-original-title="{{ __('Add new') }}">
+                                                <i class="flaticon-add"></i>
+                                            </a>
+                                        </th>
+                                    @endif
                                 </tr>
                                 </thead>
                                 <tbody>
@@ -33,14 +35,16 @@
                                             @else
                                                 <td>{{ __('Fulltime') }}</td>
                                             @endif
-                                            <td>
-                                                <a href="{{ url('admin/positions/' . $position->id . '/edit') }}" class="btn btn-warning m-btn m-btn--icon m-btn--icon-only m-btn--custom m-btn--pill" data-toggle="m-tooltip" data-placement="left" data-original-title="{{ __('Edit') }}">
-                                                    <i class="flaticon-edit-1"></i>
-                                                </a>
-                                                {!! Form::open(['url' => 'admin/positions/' . $position->id, 'method' => 'DELETE', 'class' => 'd-inline']) !!}
-                                                {!! Form::button('<i class="flaticon-cancel"></i>', ['type' => 'submit', 'class' => 'btn btn-danger m-btn m-btn--icon m-btn--icon-only m-btn--custom m-btn--pill delete', 'data-toggle' => 'm-tooltip', 'data-placement' => 'right', 'data-original-title' => __('Delete')]) !!}
-                                                {!! Form::close() !!}
-                                            </td>
+                                            @if (Auth::user()->role == config('site.permission.admin'))
+                                                <td>
+                                                    <a href="{{ url('admin/positions/' . $position->id . '/edit') }}" class="btn btn-warning m-btn m-btn--icon m-btn--icon-only m-btn--custom m-btn--pill" data-toggle="m-tooltip" data-placement="left" data-original-title="{{ __('Edit') }}">
+                                                        <i class="flaticon-edit-1"></i>
+                                                    </a>
+                                                    {!! Form::open(['url' => 'admin/positions/' . $position->id, 'method' => 'DELETE', 'class' => 'd-inline']) !!}
+                                                    {!! Form::button('<i class="flaticon-cancel"></i>', ['type' => 'submit', 'class' => 'btn btn-danger m-btn m-btn--icon m-btn--icon-only m-btn--custom m-btn--pill delete', 'data-toggle' => 'm-tooltip', 'data-placement' => 'right', 'data-original-title' => __('Delete')]) !!}
+                                                    {!! Form::close() !!}
+                                                </td>
+                                            @endif
                                         </tr>
                                     @endforeach
                                 @endif
