+++ resources/views/admin/workspace/list.blade.php
@@ -16,11 +16,13 @@
                             <th>@lang('Workspace Photo')</th>
                             <th>@lang('Workspace Name')</th>
                             <th>@lang('Total Seat')</th>
-                            <th>
-                                <a href="{{ route('workspaces.create') }}" class="btn m-btn--pill m-btn--air btn-secondary" data-toggle="m-tooltip" data-placement="left" data-original-title="@lang('Add Workspace')">
-                                    <i class="flaticon-add"></i>
-                                </a>
-                            </th>
+                            @if (Auth::user()->role == config('site.permission.admin'))
+                                <th>
+                                    <a href="{{ route('workspaces.create') }}" class="btn m-btn--pill m-btn--air btn-secondary" data-toggle="m-tooltip" data-placement="left" data-original-title="@lang('Add Workspace')">
+                                        <i class="flaticon-add"></i>
+                                    </a>
+                                </th>
+                            @endif
                         </tr>
                     </thead>
                     <tbody>
@@ -32,21 +34,23 @@
                             </td>
                             <td><h5>{{ $item->name }}</h5></td>
                             <td><h5>{{ $item->total_seat }}</h5></td>
-                            <td>
-                                <a href="{{ route('workspaces.edit', ['id' => $item->id]) }}" class="btn btn-warning m-btn m-btn--icon m-btn--icon-only m-btn--custom m-btn--pill" data-toggle="m-tooltip" data-placement="left" data-original-title="@lang('Edit Workspace')">
-                                    <i class="flaticon-edit-1"></i>
-                                </a>
-                                {!! Form::open(['route' => ['workspaces.destroy', $item->id], 
-                                    'class' => 'd-inline',
-                                    'method' => 'DELETE'
-                                ]) !!}
+                            @if (Auth::user()->role == config('site.permission.admin'))
+                                <td>
+                                    <a href="{{ route('workspaces.edit', ['id' => $item->id]) }}" class="btn btn-warning m-btn m-btn--icon m-btn--icon-only m-btn--custom m-btn--pill" data-toggle="m-tooltip" data-placement="left" data-original-title="@lang('Edit Workspace')">
+                                        <i class="flaticon-edit-1"></i>
+                                    </a>
+                                    {!! Form::open(['route' => ['workspaces.destroy', $item->id],
+                                        'class' => 'd-inline',
+                                        'method' => 'DELETE'
+                                    ]) !!}
                                     {!! Form::button('<i class="flaticon-cancel"></i>', [
                                         'class' => 'delete btn btn-danger m-btn m-btn--icon m-btn--icon-only m-btn--custom m-btn--pill',
                                         'type' => 'submit',
                                         'message' => __('Delete this item?')
                                     ]) !!}
-                                {!! Form::close() !!}
-                            </td>
+                                    {!! Form::close() !!}
+                                </td>
+                            @endif
                         </tr>
                         @empty
                             @include('admin.components.alert', ['type' => 'warning', 'message' => __('Have no data!')])
